update: log

+++ src/util/config/DefenceBaseConfig.java
@@ -1,11 +1,11 @@
 package util.config;
 
 public class DefenceBaseConfig {
-    public float minRange;
-    public float maxRange;
-    public float attackSpeed;
-    public float attackRadius;
-    public float attackArea;
-    public float attackType;
+    public double minRange;
+    public double maxRange;
+    public double attackSpeed;
+    public double attackRadius;
+    public int attackArea;
+    public int attackType;
 
 }
